// src/app/shared/model/provider.model.ts
export class ProviderModel {
    id: number;
    name: string;
    mst: string;
    mact: string;
    address: string;
    sdt: string;
    note: string;
    fax: string;
}
// CREATE TABLE `category` (
//     `id` int NOT NULL AUTO_INCREMENT,
//     `name` varchar(200),
//     `description` text(5000),
//     `imagepath` text(1000),
//     `icon` varchar(50),
//      PRIMARY KEY (`id`)
//   )
// ENGINE=InnoDB DEFAULT CHARSET=utf8;